changing schema from integer to floar

+++ datawarehouseBlackFriday/models/silver/silver_google_analytics.sql
@@ -15,6 +15,6 @@ SELECT
     CAST("activeUsers"            AS INTEGER) AS "active_users",
     CAST("sessions"               AS INTEGER) AS "sessions",
     CAST("engagedSessions"        AS INTEGER) AS "engaged_sessions",
-    CAST("averageSessionDuration" AS INTEGER) AS "average_session_duration",
+    CAST("averageSessionDuration" AS FLOAT)   AS "average_session_duration",
     CAST("screenPageViews"        AS INTEGER) AS "screen_page_views"
 FROM google_analytics
